# docs/api/radiance-openapi.json
{
  "components": {
    "schemas": {
...
      "LayoutRequest": {
        "additionalProperties": false,
        "properties": {
          "cv_enabled": {
            "default": false,
            "title": "Cv Enabled",
            "type": "boolean"
          },
          "include_svg": {
            "default": true,
            "title": "Include Svg",
            "type": "boolean"
          },
          "length_ft": {
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Length Ft",
            "type": "number"
          },
          "width_ft": {
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Width Ft",
            "type": "number"
          }
        },
        "required": [
          "length_ft",
          "width_ft"
        ],
        "title": "LayoutRequest",
        "type": "object"
      },
      "LayoutResponse": {
        "additionalProperties": false,
        "properties": {
          "all_positions": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "All Positions",
            "type": "array"
          },
          "length_ft": {
            "title": "Length Ft",
            "type": "number"
          },
          "module_count": {
            "title": "Module Count",
            "type": "integer"
          },
          "module_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Module Counts",
            "type": "object"
          },
          "png_download_name": {
            "title": "Png Download Name",
            "type": "string"
          },
          "svg": {
            "title": "Svg",
            "type": "string"
          },
          "svg_download_name": {
            "title": "Svg Download Name",
            "type": "string"
          },
          "total_cost": {
            "title": "Total Cost",
            "type": "integer"
          },
          "width_ft": {
            "title": "Width Ft",
            "type": "number"
          }
        },
        "required": [
          "length_ft",
          "width_ft",
          "module_count",
          "module_counts",
          "total_cost",
          "all_positions",
          "svg",
          "svg_download_name",
          "png_download_name"
        ],
        "title": "LayoutResponse",
        "type": "object"
      },
...
      "PublicErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "const": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error",
          "message",
          "correlation_id"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
      "RadianceRunRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "uniformity",
              "competitor",
              "visualize",
              "all",
              "metrics"
            ],
            "title": "Action",
            "type": "string"
          },
          "basis_backend": {
            "default": "rtrace",
            "enum": [
              "rtrace",
              "rcontrib_legacy",
              "rcontrib_mcpt"
            ],
            "title": "Basis Backend",
            "type": "string"
          },
          "competitor_layout": {
            "default": "full",
            "enum": [
              "full",
              "practical",
              "square",
              "rect_rect",
              "exact_tiled",
              "grid"
            ],
            "title": "Competitor Layout",
            "type": "string"
          },
          "dialux_sensor_grid": {
            "default": false,
            "title": "Dialux Sensor Grid",
            "type": "boolean"
          },
          "execution_mode": {
            "default": "precomputed",
            "enum": [
              "precomputed",
              "live_docker",
              "live_local"
            ],
            "title": "Execution Mode",
            "type": "string"
          },
          "fspm_target_ppfd_umol_m2_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fspm Target Ppfd Umol M2 S"
          },
          "fspm_target_tolerance_umol_m2_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fspm Target Tolerance Umol M2 S"
          },
          "hps_coverage_ft": {
            "default": 4.0,
            "exclusiveMinimum": 0.0,
            "maximum": 12.0,
            "title": "Hps Coverage Ft",
            "type": "number"
          },
          "hps_fixture_ppf": {
            "default": 1797.3999999999999,
            "exclusiveMinimum": 0.0,
            "maximum": 10000.0,
            "title": "Hps Fixture Ppf",
            "type": "number"
          },
          "hps_ies_variant": {
            "default": "karma",
            "maxLength": 32,
            "title": "Hps Ies Variant",
            "type": "string"
          },
          "hps_input_watts": {
            "default": 1045.0,
            "exclusiveMinimum": 0.0,
            "maximum": 2000.0,
            "title": "Hps Input Watts",
            "type": "number"
          },
          "hps_z_m": {
            "default": 0.4572,
            "exclusiveMinimum": 0.0,
            "maximum": 3.0,
            "title": "Hps Z M",
            "type": "number"
          },
          "length_ft": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Length Ft",
            "type": "number"
          },
          "match_system_ppe": {
            "default": false,
            "title": "Match System Ppe",
            "type": "boolean"
          },
          "mode": {
            "default": "SMD",
            "enum": [
              "SMD",
              "Competitor",
              "1000W HPS"
            ],
            "title": "Mode",
            "type": "string"
          },
          "mount_z_m": {
            "default": 0.4572,
            "exclusiveMinimum": 0.0,
            "maximum": 3.0,
            "title": "Mount Z M",
            "type": "number"
          },
          "overlay": {
            "default": "auto",
            "enum": [
              "auto",
              "smd",
              "spydr3",
              "hps"
            ],
            "title": "Overlay",
            "type": "string"
          },
          "peak_capping_enabled": {
            "default": false,
            "title": "Peak Capping Enabled",
            "type": "boolean"
          },
          "plant_canopy_radius_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Canopy Radius M"
          },
          "plant_columns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Columns"
          },
          "plant_growth_stage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Growth Stage"
          },
          "plant_height_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Height M"
          },
          "plant_leaf_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Leaf Count"
          },
          "plant_rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Rows"
          },
          "plant_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Seed"
          },
          "plant_spacing_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Spacing M"
          },
          "plants_enabled": {
            "default": false,
            "title": "Plants Enabled",
            "type": "boolean"
          },
          "run_basis": {
            "default": true,
            "title": "Run Basis",
            "type": "boolean"
          },
          "sim_mode": {
            "default": "standard",
            "enum": [
              "direct",
              "standard",
              "quality",
              "rigorous"
            ],
            "title": "Sim Mode",
            "type": "string"
          },
          "smd_base_ring": {
            "default": 0,
            "maximum": 16.0,
            "minimum": 0.0,
            "title": "Smd Base Ring",
            "type": "integer"
          },
          "sp_ppe": {
            "default": 2.8,
            "exclusiveMinimum": 0.0,
            "maximum": 10.0,
            "title": "Sp Ppe",
            "type": "number"
          },
          "sp_ppf": {
            "default": 2240.0,
            "exclusiveMinimum": 0.0,
            "maximum": 10000.0,
            "title": "Sp Ppf",
            "type": "number"
          },
          "sp_z_m": {
            "default": 0.4572,
            "exclusiveMinimum": 0.0,
            "maximum": 3.0,
            "title": "Sp Z M",
            "type": "number"
          },
          "subpatch_grid": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Subpatch Grid",
            "type": "integer"
          },
          "target_ppfd": {
            "default": 1000.0,
            "exclusiveMinimum": 0.0,
            "maximum": 3000.0,
            "title": "Target Ppfd",
            "type": "number"
          },
          "w_max": {
            "default": 100.0,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "W Max",
            "type": "number"
          },
          "w_min": {
            "default": 0.0,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "W Min",
            "type": "number"
          },
          "width_ft": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Width Ft",
            "type": "number"
          }
        },
        "required": [
          "action"
        ],
        "title": "RadianceRunRequest",
        "type": "object"
      },
      "RadianceRunResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact_token": {
            "title": "Artifact Token",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "outdir": {
            "title": "Outdir",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelling",
              "cancelled",
              "timed_out"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "outdir",
          "artifact_token"
        ],
        "title": "RadianceRunResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Horticulture Lighting Simulator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/layout": {
      "post": {
        "operationId": "create_layout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run Layout"
      }
    },
...
    "/radiance/run": {
      "post": {
        "operationId": "start_radiance_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RadianceRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadianceRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run Radiance"
      }
    },
...
  }
}